general: Put private stuff in extra repo

+++ ssh/install.sh
@@ -2,7 +2,9 @@
 
 echo "Installing ssh config"
 
-[ ! -d "$HOME/.ssh" ] && mkdir "$HOME/.ssh"
-[ -e "$HOME/.ssh/config" ] && rm "$HOME/.ssh/config"
+if [ -e "./config" ]; then
+    [ ! -d "$HOME/.ssh" ] && mkdir "$HOME/.ssh"
+    [ -e "$HOME/.ssh/config" ] && rm "$HOME/.ssh/config"
 
-ln -s "$HOME/dotfiles/ssh/config" "$HOME/.ssh/config"
+    ln -s "$HOME/dotfiles/ssh/config" "$HOME/.ssh/config"
+fi
